This is a challenge from r/dailyprogrammer. It is writting in Java.
The input is a single number: the number at which the game starts. Write
a program that plays the Threes game, and outputs a valid sequence of
steps you need to take to get to 1. Each step should be output as the
number you start at, followed by either -1 or 1 (if you are
adding/subtracting 1 before dividing), or 0 (if you are just dividing).
The last line should simply be 1.